+++ R/catalog_makecluster.r
@@ -44,38 +44,23 @@ catalog_makecluster = function(ctg, res, start = c(0,0), plot = TRUE)
   }
   else
   {
-    start = start %% res
-
-    # dimension of the clusters (width = height)
-    # rounded up to a multiple of the resolution
+    # Dimension of the clusters (width = height) rounded up to a multiple of the resolution
     width = ceiling(size/res) * res
 
     # Bounding box of the catalog
     bbox = with(ctg@data, c(min(`Min X`), min(`Min Y`), max(`Max X`), max(`Max Y`)))
 
-    # Buffer around the bbox as a multiple of the resolution
-    # This enables to start and end clusters at exact mutilples of the resolution.
-    buffered_bbox = bbox + c(-res, -res, +res, +res)
-    buffered_bbox = round_any(buffered_bbox, res)
-    buffered_bbox = buffered_bbox + c(-res, -res, +res, +res)
+    # Shift to align the grid
+    shift = numeric(2)
+    shift[1] = (bbox[1] - start[1]) %% width
+    shift[2] = (bbox[2] - start[2]) %% width
 
-    if (!all((buffered_bbox %% res) == 0))
-      stop("Internal error, please report the error to the maintainer: bounding box incorrect.")
-
-    # Shift the bounding box to match with the start parameter (grid_metrics)
-    buffered_bbox = buffered_bbox + c(start[1], start[2], start[1], start[2])
-
-    # Generate coordinates of clusters
-    xmin = seq(buffered_bbox[1], buffered_bbox[3], width)
-    ymin = seq(buffered_bbox[2], buffered_bbox[4], width)
-
-    xmin = xmin[xmin < max(ctg@data$`Max X`)]
-    ymin = ymin[ymin < max(ctg@data$`Max Y`)]
-
-    X = expand.grid(xmin = xmin, ymin = ymin)
-
-    xmin = X$xmin
-    ymin = X$ymin
+    # Generate coordinates of bottom left clusters corners
+    xmin = seq(bbox[1] - shift[1], bbox[3], width)
+    ymin = seq(bbox[2] - shift[2], bbox[4], width)
+    grid = expand.grid(xmin = xmin, ymin = ymin)
+    xmin = grid$xmin
+    ymin = grid$ymin
     xmax = xmin + width
     ymax = ymin + width
   }